fix: address Copilot review on stats drill-in PR
Trim the drill_in module doc to 5 lines, move its ~500-line inline
test suite to a sibling drill_in/tests.rs per the repo's testing
convention, and assert the trend chart renders in the drill-in
open/close Playwright test (AC2 was only partially covered).
